+++ wappstoiot/connections/sslsocket.py
@@ -11,7 +11,7 @@ from typing import Callable
 from typing import Optional
 from typing import Union
 
-from .protocol import Status
+from .protocol import StatusID
 from .protocol import Connection
 
 
@@ -30,7 +30,7 @@ class TlsSocket(Connection):
 
         self.observer_name = "CONNECTION"
         self.observer = observer if observer else lambda st, nd: None
-        self.observer.post(Status.DISCONNETCED, None)
+        self.observer.post(StatusID.DISCONNETCED, None)
 
         self.send_ready = threading.Lock()
 
@@ -154,6 +154,8 @@ class TlsSocket(Connection):
             self.log.exception(msg)
             # UNSURE: How do we hit this?
             return False
+        except AttributeError:
+            return False
         else:
             self.log.debug(f"Raw Data Send: {data}")
             return True
@@ -178,7 +180,7 @@ class TlsSocket(Connection):
         while self.socket:
             try:
                 data_chunk = self.socket.recv(self.RECEIVE_SIZE)
-            except socket.timeout:
+            except (socket.timeout, TimeoutError):
                 continue
             except OSError:
                 # UNSURE:
@@ -199,30 +201,29 @@ class TlsSocket(Connection):
                 self.log.debug(f"Raw Data Received: {data}")
                 return parsed_data
 
-    def connect(self) -> None:
+    def connect(self) -> Optional[bool]:
         """
         Connect to the server.
 
         Attempts a connection to the server on the provided addres and port.
 
         Returns:
-            'True' if the connection was successful else
-            'False'
+            'True' if the connection was successful.
         """
 
         try:
             self.log.info("Trying to Connect.")
-            self.observer.post(Status.CONNECTING, None)
+            self.observer.post(StatusID.CONNECTING, None)
             # self.socket.settimeout(10)  # Why?
             self.socket.connect((self.address, self.port))
             # self.socket.settimeout(None)  # Why?
             self.log.info(
                 f"Connected on interface: {self.socket.getsockname()[0]}"
             )
-            self.observer.post(Status.CONNECTED, None)
+            self.observer.post(StatusID.CONNECTED, None)
             # if self.sockt_thread is None:
             #     self._start()
-            # return True
+            return True
 
         except Exception as e:
             self.log.error("Failed to connect: {}".format(e))
@@ -250,8 +251,11 @@ class TlsSocket(Connection):
                 retry_limit -= 1
             self.disconnect()
             self._socket_setup()
-            if self.connect():
-                return True
+            try:
+                if self.connect():
+                    return True
+            except OSError:
+                pass  # NOTE: Happens if it have forgotten the IP for the url.
             self.log.info("Trying to reconnect in 5 seconds")
             time.sleep(5)
         return False
@@ -268,12 +272,12 @@ class TlsSocket(Connection):
         Closes the socket object connection.
         """
         self.log.info("Closing connection...")
-        self.observer.post(Status.DISCONNECTING, None)
+        self.observer.post(StatusID.DISCONNECTING, None)
         if self.socket:
             self.socket.close()
             self.socket = None
         if self.raw_socket:
             self.raw_socket.close()
             self.raw_socket = None
-        self.observer.post(Status.DISCONNETCED, None)
+        self.observer.post(StatusID.DISCONNETCED, None)
         self.log.info("Connection closed!")